requirements.puerta.txt
# install opencv like this:
# first install the c++ binary: sudo apt-get install libopencv-dev
# then install python bindings: sudo apt-get install python3-opencv
requests
python-dotenv
evdev
RPi.GPIO
pytest
smbus2
pyserial
pyzmq
setproctitle
#systemd-python  # sudo apt install libsystemd-dev
rpi-lcd
unidecode
sentry_sdk